Moved tests to verify, Added SvgIcon and svg imports for FA Free

+++ src/Blazor.FontAwesome6/Vector/SvgTransform.cs
@@ -3,10 +3,10 @@
 public record SvgTransform
 {
     public static SvgTransform None { get; } = new SvgTransform();
-    public int Size { get; init; } = 16;
-    public int X { get; init; } = 0;
-    public int Y { get; set; } = 0;
-    public int Rotate { get; init; } = 0;
+    public double Size { get; init; } = 16;
+    public double X { get; init; } = 0;
+    public double Y { get; set; } = 0;
+    public double Rotate { get; init; } = 0;
     public bool FlipX { get; init; } = false;
     public bool FlipY { get; init; } = false;
 }
